+++ src/ngraph/frontend/onnx_import/utils/reshape.cpp
@@ -23,9 +23,9 @@
 #include <vector>
 
 #include "exceptions.hpp"
+#include "ngraph/builder/reshape.hpp"
 #include "ngraph/op/reshape.hpp"
 #include "ngraph/op/slice.hpp"
-#include "ngraph/op/util/reshape.hpp"
 #include "utils/common.hpp"
 #include "utils/reshape.hpp"
 
@@ -111,7 +111,7 @@ namespace ngraph
                         output_shape.push_back(axis);
                     }
                 }
-                return ngraph::op::util::reshape(node, output_shape);
+                return ngraph::builder::reshape(node, output_shape);
             }
 
             std::shared_ptr<ngraph::Node> collapse(const std::shared_ptr<ngraph::Node>& node,
@@ -129,7 +129,7 @@ namespace ngraph
                 output_shape.insert(std::end(output_shape),
                                     std::next(std::begin(shape), end_axis + 1),
                                     std::end(shape));
-                return ngraph::op::util::reshape(node, output_shape);
+                return ngraph::builder::reshape(node, output_shape);
             }
 
             std::shared_ptr<ngraph::Node> expand_dims(const std::shared_ptr<ngraph::Node>& node,
@@ -159,7 +159,7 @@ namespace ngraph
                              "Scalar value can't be derived from a node with ",
                              node_shape);
 
-                return ngraph::op::util::reshape(node, Shape{});
+                return ngraph::builder::reshape(node, Shape{});
             }
 
         } // namespace  reshape